+++ lib/ig_markets/cli/activities_command.rb
@@ -9,24 +9,9 @@ module IGMarkets
       def activities
         self.class.begin_session(options) do |dealing_platform|
           dealing_platform.account.recent_activities(seconds).each do |activity|
-            print_activity activity
+            Output.print_activity activity
           end
         end
-      end
-
-      private
-
-      def seconds
-        (options[:days].to_f * 60 * 60 * 24).to_i
-      end
-
-      def print_activity(activity)
-        puts <<-END
-#{activity.deal_id}: \
-#{activity.size} of #{activity.epic}, \
-level: #{activity.level}, \
-result: #{activity.result}
-END
       end
     end
   end