+++ lib/ig_markets/cli/watchlists_command.rb
@@ -7,26 +7,16 @@ module IGMarkets
       def watchlists
         self.class.begin_session(options) do |dealing_platform|
           dealing_platform.watchlists.all.each do |watchlist|
-            print_watchlist watchlist
+            Output.print_watchlist watchlist
 
-            watchlist.markets.each do |market|
+            watchlist.markets.each do |market_overview|
               print '  - '
-              print_market_overview market
+              Output.print_market_overview market_overview
             end
+
             print "\n"
           end
         end
-      end
-
-      private
-
-      def print_watchlist(watchlist)
-        puts <<-END
-#{watchlist.id}: #{watchlist.name}, \
-editable: #{watchlist.editable}, \
-deleteable: #{watchlist.deleteable}, \
-default: #{watchlist.default_system_watchlist}
-END
       end
     end
   end